+++ src/app/locales/en/components/content-atm-inventories-lambdas/atm-lambda-form.js
@@ -10,10 +10,6 @@ const argumentAndResultTypes = {
   object: {
     label: 'Object',
   },
-  // TODO: VFS-7816 uncomment or remove future code
-  // histogram: {
-  //   label: 'Histogram',
-  // },
   anyFile: {
     label: 'Any file',
   },
@@ -56,6 +52,9 @@ const argumentAndResultTypes = {
   },
   onedatafsCredentials: {
     label: 'OnedataFS credentials',
+  },
+  timeSeriesMeasurement: {
+    label: 'Time series measurement',
   },
 };